♻️ Break up summary reporter into dedicated plugin
