+++ lib/plugin/reporter.js
@@ -17,7 +17,7 @@ export class ReporterResolver {
 
     test.on('test:prepare', async () => {
       if (!this.#reporters.size && !this.#remote)
-        await this.#configure('spec');
+        await this.#configure(['spec', 'summary']);
     });
   }
