+++ lib/reporter/dot.js
@@ -42,13 +42,8 @@ export class DotReporter extends Reporter {
         if (newline) count = 0;
       }
 
-      if (type === 'test:end') {
+      if (type === 'test:end')
         yield '\n';
-
-        yield this.summarize(data, {
-          colors: this.#colors
-        });
-      }
     }
 
     yield '\n';
