+++ lib/reporter/index.js
@@ -1,4 +1,5 @@
 export { Reporter } from './base.js';
 export { DotReporter, dotReporter } from './dot.js';
 export { SpecReporter, specReporter } from './spec.js';
+export { SummaryReporter, summaryReporter } from './summary.js';
 export { TapReporter, tapReporter } from './tap.js';
