+++ lib/reporter/spec.js
@@ -8,16 +8,13 @@ export class SpecReporter extends Reporter {
     'test:start': '🚀 ',
     'test:skip': '💤 ',
     'test:pass': '✅ ',
-    'test:fail': '❌ ',
-    'test:summary': '🏁 ',
-    'test:remaining': '🚧 '
+    'test:fail': '❌ '
   };
 
   #colors = {
     'test:skip': 'blue',
     'test:pass': 'green',
-    'test:fail': 'red',
-    'test:remaining': 'yellow'
+    'test:fail': 'red'
   };
 
   configure(config) {
@@ -62,13 +59,6 @@ export class SpecReporter extends Reporter {
         let cache = grouped.get(data.origin) ?? [];
         grouped.set(data.origin, cache);
         cache.push(data);
-      }
-
-      if (type === 'test:end') {
-        yield this.summarize(data, {
-          symbols: this.#symbols,
-          colors: this.#colors
-        });
       }
     }
